buy and sell alerts, email/text + in application

+++ QuickFlip/Controllers/BuyController.cs
@@ -281,7 +281,8 @@ namespace QuickFlip.Controllers
             // delete old new offer and outbid for this post
             foreach (var alert in oldAlerts)
             {
-                if (alert.Type == AlertType.NewOffer || alert.Type == AlertType.Outbid)
+                if (alert.Type == AlertType.NewOffer || 
+                    (alert.Type == AlertType.Outbid && alert.UserId == newOffer.UserId))
                 {
                     BusinessLogic.DeleteAlert(alert.AlertId); 
                 }
@@ -301,7 +302,7 @@ namespace QuickFlip.Controllers
             // make new offer
             newOffer = BusinessLogic.CreateOffer(newOffer);
 
-            // send pos owner an alert saying a new offer arrived
+            // send post owner an alert saying a new offer arrived
             BusinessLogic.CreateAlert(postId, post.UserId, newOffer.OfferId, AlertType.NewOffer);
 
             return RedirectToAction("Index", new { id = (CommunityAbbrev)post.CommunityId });
